requirements.txt
PyMuPDF
pandas
python-docx
openpyxl
tiktoken